pyproject.toml
[project]
name = "finance-discord-bot"
version = "0.1.0"
description = "Discord bot for financial data analysis, integrating with various API services"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.3.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.0.0",
    "apscheduler>=3.10.0",
    "feedparser>=6.0.12",
    "openai>=2.14.0",
]
